Correcion de detalles pre-compilado

+++ ProyectoparqueaderoPS2025/Repos/libr_dominio/Entidades/Tarifas.cs
@@ -6,7 +6,7 @@ namespace libr_dominio.Entidades
     {
         public int Id { get; set; }
         public string? Tipo { get; set; }
-        public double Valor { get; set; }
+        public Decimal Valor { get; set; }
 
         [ForeignKey("ParqueaderosClientes")] public List<ParqueaderosClientes>? _ParqueaderosClientes { get; set; }
     }
